line 159: self.layers[0].h is initialized as of the same type as X_i
Former-commit-id: b0980a4bd9a37761b19423092e279d90f1f9a258
Former-commit-id: 3f3d915bdacbb9983972ce483d5f5966d7c1ad3d

+++ easysurrogate/methods/NN.py
@@ -156,7 +156,7 @@ class ANN:
         if self.bias == False:
             self.layers[0].h = X_i.T
         else:
-            self.layers[0].h = xp.ones([self.n_in + 1, batch_size])
+            self.layers[0].h = xp.ones([self.n_in + 1, batch_size],dtype=X_i.dtype)
             self.layers[0].h[0:self.n_in, :] = X_i.T
                     
         for i in range(1, self.n_layers+1):
